requirement.txt
Pillow==9.4.0
pillow-avif-plugin==1.3.1
img2pdf==0.4.4
PyPDF2[full]
